Final update!
The moment you've been waiting for, since the first press conference in the summer, is here! 
I present to you the iPod... ehm no not the iPod, more like an mp3 player that 'maybe' kind of resembles an iPod (don't sue me Apple please). 
Anyway it's working, I think I'm happy with it and it renewed some childhood memories for me as I had an iPod... I mean some mp3 player (don't sue me Apple) as a kid that looked like this and I loved the hell out of it. 
So this project is my love letter to this particular mp3 player :)

I hope it's at least fine, it cost me some nerves and made me grow grey hair